// src/services/api/instansi/akademik/kelas/sesi-pembelajaran/list.ts
import { makeJwtGetRequestTableApi } from '@/utils/api'

export type DataType = {
  id: string
  id_kelas: string
  judul: string
  pertemuan: number
  hari:
    | 'Senin'
    | 'Selasa'
    | 'Rabu'
    | 'Kamis'
    | 'Jumat'
    | 'Sabtu'
    | 'Minggu'
    | null
  waktu_mulai: string
  waktu_sampai: string
  tanggal_realisasi: string | null
  status: 'Sedang Berlangsung' | 'Belum Dibuka' | 'Telah Berakhir'
  jenis_absensi_pengajar: 'GPS' | 'Swafoto' | 'GPS dan Swafoto' | 'QR Code'
  jenis_absensi_peserta:
    | 'Manual'
    | 'Otomatis'
    | 'GPS'
    | 'Swafoto'
    | 'GPS dan Swafoto'
    | 'QR Code'
  status_absensi_pengajar: 'Hadir' | 'Izin' | 'Sakit' | 'Alpha' | null
  latitude: number | null
  longitude: number | null
  swafoto: string | null
  id_ruangan: string | null
  created_at: string
  updated_at: string
  nama_pengajar: string
  swafoto_url: string | null
  lokasi_pertemuan: string
  total_bahan_ajar: number
  status_absensi_siswa: 'Hadir' | 'Izin' | 'Sakit' | 'Alpha' | null
}

export const listSesiPembelajaranApi = async ({
  jwt,
  idKelas,
  page = 1,
  search = '',
  perPage = 20,
  sortBy,
  order,
}: {
  jwt: string
  idKelas: string
  page?: number
  search?: string
  perPage?: number
  sortBy?: string
  order?: 'asc' | 'desc'
}) =>
  makeJwtGetRequestTableApi<DataType>(
    `${process.env.NEXT_PUBLIC_API_URL}/instansi/kelas/${idKelas}/pertemuan`,
    jwt,
    {
      current_page: page,
      keyword: search,
      per_page: perPage,
      sort_by: sortBy,
      order: order,
    }
  )